moved autocorrect into separate file
